Avoid sparse dict compaction over filters
