+++ vortex-array/src/arrays/dict/vtable/cardinality.rs
@@ -10,8 +10,6 @@
 use vortex_mask::Mask;
 
 use crate::arrays::PrimitiveArray;
-use crate::dtype::IntegerPType;
-
 const SAMPLE_SIZE: usize = 128;
 const REPEATED_CODE_PROBE_SIZE: usize = 16;
 
@@ -21,10 +19,7 @@ const REPEATED_CODE_PROBE_SIZE: usize = 16;
 /// next, cheaper filter for cases that are not sparse by row count alone: dense dictionaries should
 /// not pay the full estimator cost unless the code stream first shows evidence of repeated codes.
 /// A `true` result only means "run the estimator"; it is not enough to compact by itself.
-pub(super) fn has_repeated_code_sample<P: IntegerPType>(
-    codes: &PrimitiveArray,
-    validity_mask: &Mask,
-) -> bool {
+pub(super) fn has_repeated_code_sample(codes: &PrimitiveArray, validity_mask: &Mask) -> bool {
     let sample_count = codes.len().min(REPEATED_CODE_PROBE_SIZE);
     let mut observed_codes = Vec::<usize>::with_capacity(sample_count);
 
@@ -34,7 +29,7 @@ pub(super) fn has_repeated_code_sample<P: IntegerPType>(
             continue;
         }
 
-        let code = codes.as_slice::<P>()[idx].as_();
+        let code = code_at(codes, idx);
         if observed_codes.contains(&code) {
             return true;
         }
@@ -49,7 +44,7 @@ pub(super) fn has_repeated_code_sample<P: IntegerPType>(
 /// The estimator samples deterministic bucket midpoints so repeated executions make the same
 /// compaction decision for the same input. Returning `None` means no valid sampled codes were seen.
 /// A returned value should only be used to decide whether an exact pass is worth attempting.
-pub(super) fn estimate_code_cardinality<P: IntegerPType>(
+pub(super) fn estimate_code_cardinality(
     codes: &PrimitiveArray,
     validity_mask: &Mask,
 ) -> Option<usize> {
@@ -64,7 +59,7 @@ pub(super) fn estimate_code_cardinality<P: IntegerPType>(
             continue;
         }
 
-        let code = codes.as_slice::<P>()[idx].as_();
+        let code = code_at(codes, idx);
         if let Some((_, count)) = observed_codes
             .iter_mut()
             .find(|(observed, _)| *observed == code)
@@ -126,6 +121,10 @@ fn sample_index(sample_idx: usize, len: usize, sample_count: usize) -> usize {
     let bucket_end = (sample_idx + 1) * len / sample_count;
 
     ((bucket_start + bucket_end) / 2).min(len - 1) as usize
+}
+
+fn code_at(codes: &PrimitiveArray, idx: usize) -> usize {
+    usize::try_from(codes.as_slice::<u64>()[idx]).unwrap_or(usize::MAX)
 }
 
 fn div_ceil(numerator: usize, denominator: usize) -> usize {
